move utils to common and paint to svgcanvas
