+++ src/editor/components/jgraduate/Slider.js
@@ -1,4 +1,4 @@
-import { findPos } from './Util.js'
+import { findPos } from '../../../common/util.js'
 /**
  * Whether a value is `null` or `undefined`.
  * @param {any} val
