+++ src/editor/components/jgraduate/jQuery.jPicker.js
@@ -18,7 +18,7 @@
 /* eslint-disable max-len */
 import ColorValuePicker from './ColorValuePicker.js'
 import Slider from './Slider.js'
-import { findPos, mergeDeep } from './Util.js'
+import { findPos, mergeDeep } from '../../../common/util.js'
 
 /**
 * @external Math
